Dev (#231)
* Add files via upload

* Add files via upload

* docs: uploading updated document

* add Planejamento da Sprint 2

* remoção de diagramas desnecessário

* Organizando pasta de planejamento

* Add files via upload

* Add files via upload

* Add files via upload

* Add planning artifacts for Sprint 2, Sprint 3, and multiple versions of Sprint 4

* Adding diagram files in master (they were in dev)

* moving diagram between packages and deleting not updated diagrams

* adicionando diagrama de classes em pdf

* Add files via upload

* Add files via upload

* feat: adicionando o planejamento da Sprint 5

* adicionando diagrama

* adicionando diagrama de casos de uso

* adicionando diagrama de entidade relacionamento

* adicionando diagrama de classe

* Add files via upload

* Remove notification icon and badge from header component

* Add file upload configuration to application properties

* Refactor invoice and combined score services to streamline client lookup and date handling; remove unused environment variable from properties

* Update database configuration in application.properties; adjust MySQL connection settings and enable HikariCP properties

* Update application.properties for local database config

Uncommented local database configuration properties.

* fix: tipo do atributo quantidade

---------

+++ Codigo/Back/src/main/java/com/hortifruti/sl/hortifruti/dto/purchase/UpdateInvoiceProduct.java
@@ -3,4 +3,4 @@ package com.hortifruti.sl.hortifruti.dto.purchase;
 import java.math.BigDecimal;
 
 public record UpdateInvoiceProduct(
-    String code, String name, BigDecimal price, Integer quantity, String unitType) {}
+    String code, String name, BigDecimal price, BigDecimal quantity, String unitType) {}
